prepare main() for separation from lib portion of code
switched flags to XXXVar(pointer) so the dereferences aren't necessary
